requirements.txt
django
google
google-api-python-client
google-cloud
google-generativeai
google-oauth
jupyter
numpy
pandas
psycopg
psycopg2-binary
pydantic
python-dotenv
uuid
uvicorn
vertexai